# Conflicts:
#	CMakeLists.txt
#	datastructures/polygon.cpp
#	datastructures/polygon.h
#	datastructures/quadtree.cpp
#	datastructures/quadtree.h
#	main.cpp

+++ datastructures/polygon.h
@@ -36,7 +36,7 @@ public:
 
     lineseg getEdge(int index) const;
 
-    double area() const;
+    long double area() const;
 
     bool isConvex() const;
